requirements.txt
flask==3.0.0
groq>=0.9.0
openpyxl==3.1.5
python-dotenv
python-docx>=0.8.11